Seperated seperate classes into seperate files, to better control and debug data in future.
