+++ many7seriesdata.py
@@ -1,267 +1,7 @@
 from swiftBaseClass import swiftBaseClass as sbc
 import random
 import pandas as pd
-from datetime import datetime, timedelta
-
-class many103data(sbc):
-    def exposure_details(self):
-        uumid, amt, ccy, mt, vdate, corr = super().exposure_details()
-        mt=random.choice([keys for keys,values in self.EXPOSURE_MT.items() if values=='GROUP1'])
-        return uumid, amt, ccy, mt, vdate, corr
-    
-    def generate_records(self):
-        swift_records, pp_records, cbs_records = [], [], []
-
-        # Happy Case Scenario, Many/one CBS entry sum(cbsamount)>=swift amount=ppamount(Reconciled Case)
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # Swift Not present, Many/one CBS entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # PP Not present, Many/one CBS entry, Swift Present
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # PP 2 entries, Swift 1 entry, CBS 1/many entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) #Adjust Time/Money
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # PP 2 entries, Swift Not present, Many/one CBS entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) #Adjust Time/Money
-
-        # Swift 2 entry, PP 1 entry, CBS 1/many entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # Alter Time
-
-        # Swift 2 entry, PP no entry, CBS 1/many entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # Alter Time
-
-        # Swift 2 entry, PP 2 entry, CBS 1/many entry
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            cbs_random_split=random.randint(1,3)
-
-            if cbs_random_split!=1:
-                amt=int(amt)
-
-                cut_points = sorted(random.sample(range(1, amt), cbs_random_split - 1))
-                cut_points = [0] + cut_points + [amt]
-                amt_split = [cut_points[i+1] - cut_points[i] for i in range(len(cut_points)-1)]
-
-                for split_amt in amt_split:
-                    cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            else :
-                cbs_records.append(self.make_cbs(uumid, split_amt, ccy, vdate, "Scenario 10", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # Alter time
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # Alter Time
-
-        
-        # Swift 1 entry, PP 1 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # Swift no entry, PP 1 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-            
-        # Swift 1 entry, PP no entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # Swift 1 entry, PP 2 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # adjust time
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-
-        # Swift no entry, PP 2 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # adjust time
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-        
-        # Swift 2 entry, PP 1 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) #Adjust Time
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-
-        # Swift 2 entry, PP no entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) #Adjust Time
-
-        # Swift 2 entry, PP 2 entry, No CBS entries
-        for i in range(10):
-            uumid, amt, ccy, mt, vdate, corr = self.exposure_details() 
-            
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S")))
-            swift_records.append(self.make_swift(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) #Adjust Time
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) 
-            pp_records.append(self.make_pp(uumid, amt, ccy, mt, vdate, corr, "Scenario 2", datetime.now().strftime("%d-%m-%Y %H:%M:%S"))) # adjust time
-
-
-        return pd.DataFrame(cbs_records), pd.DataFrame(swift_records), pd.DataFrame(pp_records)
-
-
-
-
-
-
-
-
-
-
+from datetime import datetime
 
 class many7seriesdata(sbc):
     def exposure_details(self):
@@ -537,25 +277,3 @@ class many7seriesdata(sbc):
 
 
         return pd.DataFrame(cbs_records), pd.DataFrame(swift_records), pd.DataFrame(pp_records)
-    
-
-if __name__ == "__main__":
-    happygenerator = sbc()
-    many103generator = many103data()
-    many7XXgenerator = many7seriesdata()
-
-    cbs_df1, swift_df1, pp_df1 = happygenerator.generate_records()
-    cbs_df2, swift_df2, pp_df2 = many103generator.generate_records()
-    cbs_df3, swift_df3, pp_df3 = many7XXgenerator.generate_records()
-
-    cbs_df = pd.concat([cbs_df1, cbs_df2, cbs_df3], ignore_index=True)
-    swift_df = pd.concat([swift_df1, swift_df2, swift_df3], ignore_index=True)
-    pp_df = pd.concat([pp_df1, pp_df2, pp_df3], ignore_index=True)
-
-    cbs_file_name, swift_file_name, pp_file_name = happygenerator.generate_file_names()
-
-    cbs_df.to_csv(cbs_file_name, index=False, sep="|")
-    pp_df.to_csv(pp_file_name, index=False, sep="|")
-    swift_df.to_excel(swift_file_name, index=False)
-
-    print("Files generated successfully:", cbs_file_name, swift_file_name, pp_file_name)
